usb: dwc3: Prevent continuous retries on error event
Upon USB core generating an erratic error event, the
dwc3-msm driver resets and restarts the core. Even
after this if core could not recover, the driver
keep on trying the same in a infinite loop. This could
lead to the hogging of system resources leading to
system failure. Prevent this by trying a maximum of three
times to recover upon erratic error event.

An example of continuous generation of error events is
a DCP charger detected as SDP by the PMIC and core entering
device mode. In this case, the core could not recover
even after several tries.

Change-Id: Ic2f2f10e6bcc7480be5116746072b0fb1651ce10
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>